Move views logic from tmp directory to Gatsby page
