+++ src/pages/index.tsx
@@ -1,19 +1,20 @@
 import * as React from "react"
-
+import { NavBar } from '../shared/components'
 import { withAuthenticator } from '@aws-amplify/ui-react'
 
 import Amplify from 'aws-amplify'
 import awsmobile from '../aws-exports'
-import HomePage from "../tmp/Home";
 
 Amplify.configure(awsmobile)
 
-const IndexPage = () => {
-  return (
-      <div>
-        <HomePage />
-      </div>
-  )
+const HomePage = () => {
+    return (
+        <div>
+            <NavBar />
+        </div>
+    )
 }
 
-export default withAuthenticator(IndexPage)
+export default withAuthenticator(HomePage)
+
+
