+++ src/pages/webinars.tsx
@@ -1,10 +1,18 @@
 import * as React from "react"
-import {SearchPage} from "../tmp";
 import {withAuthenticator} from "@aws-amplify/ui-react";
 import Amplify from "aws-amplify";
 
 import awsmobile from "../aws-exports";
+import {NavBar} from "../shared/components";
 Amplify.configure(awsmobile)
 
+const SearchPage = () => {
+    return (
+        <div>
+            <NavBar />
+            <p>Coming Soon</p>
+        </div>
+    )
+}
 
 export default withAuthenticator(SearchPage)
