--- src/tmp/Home.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import { NavBar } from '../shared/components'
-
-const HomePage = () => {
-    return (
-        <div>
-            <NavBar />
-        </div>
-    )
-}
-
-export default HomePage
