--- src/tmp/LiveStream.tsx
@@ -1,13 +1,0 @@
-import * as React from "react"
-import { NavBar } from '../shared/components'
-
-const LiveApp = () => {
-    return (
-        <div>
-            <NavBar />
-            <p>Coming Soon</p>
-        </div>
-    )
-}
-
-export default LiveApp
