--- src/tmp/Search.tsx
@@ -1,13 +1,0 @@
-import * as React from "react"
-import { NavBar } from '../shared/components'
-
-const SearchPage = () => {
-    return (
-        <div>
-            <NavBar />
-            <p>Coming Soon</p>
-        </div>
-    )
-}
-
-export default SearchPage
