--- src/tmp/Webinars.tsx
@@ -1,12 +1,0 @@
-import { NavBar } from '../shared/components'
-
-const WebinarsApp = () => {
-    return (
-        <div>
-            <NavBar />
-            <p>Coming Soon</p>
-        </div>
-    )
-}
-
-export default WebinarsApp
