+++ src/tmp/index.ts
@@ -1,18 +1,8 @@
 import Dashboard from './Dashboard/Dashboard'
-import VodApp from './VideoOnDemand'
 import VideoPage from './VideoPlayer/VideoPage'
-import HomePage from './Home'
-import LiveApp from './LiveStream'
-import WebinarsApp from './Webinars'
-import SearchPage from './Search'
 
 export {
     Dashboard,
-    VodApp,
     VideoPage,
-    HomePage,
-    LiveApp,
-    WebinarsApp,
-    SearchPage,
 }
